subdivide device objects
